# Conflicts:
#	backend/accounting-service/src/jvmMain/kotlin/services/Wallets.kt
#	backend/app-aau-service/build.gradle.kts
#	backend/app-aau-service/src/jvmMain/kotlin/rpc/ComputeController.kt
#	backend/app-kubernetes-service/api/src/commonMain/kotlin/api/KubernetesIPMaintenance.kt
#	backend/app-kubernetes-service/src/jvmMain/kotlin/services/NetworkIPService.kt
#	backend/app-kubernetes-service/src/jvmMain/resources/db/migration/V7__IP_Remapping.sql
#	backend/audit-ingestion-service/src/jvmMain/kotlin/audit/ingestion/processors/AuditProcessor.kt
#	backend/auth-service/api/src/commonMain/kotlin/UserDescriptions.kt
#	backend/elastic-management-service/build.gradle.kts
#	backend/grant-service/api/src/commonMain/kotlin/Grants.kt
#	backend/grant-service/src/jvmMain/kotlin/grant/rpc/GrantController.kt
#	backend/grant-service/src/jvmMain/kotlin/grant/services/Applications.kt
#	backend/indexing-service/build.gradle.kts
#	backend/project-service/api/src/commonMain/kotlin/Project.kt
#	backend/project-service/api/src/commonMain/kotlin/Projects.kt
#	backend/project-service/src/jvmMain/kotlin/project/services/QueryService.kt
#	backend/project-service/src/jvmTest/kotlin/project/services/QueryServiceTest.kt
#	backend/storage-service/build.gradle.kts
#	backend/storage-service/src/jvmMain/kotlin/file/services/BulkUploadService.kt
#	frontend-web/webclient/app/Applications/AppToolLogo.tsx
#	frontend-web/webclient/app/Applications/Jobs/View.tsx
#	frontend-web/webclient/app/UCloud/index.ts
#	frontend-web/webclient/app/UtilityFunctions.tsx

+++ backend/app-orchestrator-service/src/jvmMain/resources/db/migration/V48__JobMigration.sql
@@ -164,7 +164,7 @@ begin
                     job.application_name,
                     job.application_version,
                     job.reserved_price_per_unit,
-                    (job.max_time_hours * 3600 * 1000) + (job.max_time_minutes * 60 * 1000) +
+                    (job.max_time_hours::bigint * 3600 * 1000) + (job.max_time_minutes * 60 * 1000) +
                     (job.max_time_seconds * 1000),
                     coalesce(job.credits_charged, 0),
                     job.reserved_provider,
